ubah tipe data hitung ke double

+++ src/Hitung.java
@@ -1,3 +1,3 @@
 public interface Hitung {
-    public double hitung(int angkanya);
+    public double hitung(double angkanya);
 }
